+++ packages/kobber-components/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @gyldendal/kobber-components
 
+## 0.3.8
+
+### Patch Changes
+
+- Progress bar updates
+- Updated dependencies
+  - @gyldendal/kobber-base@0.3.8
+
+## 0.3.7
+
+### Patch Changes
+
+- Updated default grid paddings
+- Updated dependencies
+  - @gyldendal/kobber-base@0.3.7
+
 ## 0.3.6
 
 ### Patch Changes